fixes for g++ 3
git-svn-id: svn+ssh://svn.code.sf.net/p/snowstorm/code/trunk/openglad@425 636a839e-d62e-0410-834d-a1d0aacdb804

+++ src/treasure.cpp
@@ -291,9 +291,9 @@ short treasure::eat_me(walker  * eater)
 			death();
 			return 1;
 		case FAMILY_KEY: // get the key to this door ..
-			if (!(eater->keys & (long)(pow(2, stats->level)) )) // just got it?
+			if (!(eater->keys & (long)(pow((double) 2, stats->level)) )) // just got it?
 			{
-				eater->keys |= (long) (pow(2, stats->level)); // ie, 2, 4, 8, 16...
+				eater->keys |= (long) (pow((double)2, stats->level)); // ie, 2, 4, 8, 16...
 				if (eater->myguy)
 					sprintf(message, "%s picks up key %d", eater->myguy->name,
 					        stats->level);
